+++ src/components/Users.js
@@ -1,18 +1,12 @@
-import React, { useEffect } from 'react';
-import { useDispatch, useSelector } from 'react-redux';
+import { useSelector } from 'react-redux';
 
 import User from './User';
 
 const Users = () => {
 
-  const dispatch = useDispatch();
-
   let users = useSelector(state => state.users);
   console.log('My users: ', users)
 
-  useEffect(() => {
-    dispatch({ type: 'FETCH_USERS'})
-  }, [dispatch]);
   
   return (
     <div className="col-5 m-3">
